Fix login without Redux
Sadly, had to make a big change. But it was worth it.

The React app is now divided neatly into folder according to Cory
House's guidelines. Still a lot of cruft coming from almost-using-Redux,
but I'll remove it soon.

Still getting a warning for "Can't perform a React state update on an
unmounted component" but will look into that later.